Ricky: moved exceptions to common package
