+++ src/main/java/com/epsilon/homebanking/common/exception/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package com.epsilon.homebanking.exception;
+package com.epsilon.homebanking.common.exception;
 
 public class InvalidCredentialsException extends
     RuntimeException {
