+++ src/main/java/com/epsilon/homebanking/security/delegate/AuthenticationDelegate.java
@@ -1,10 +1,9 @@
 package com.epsilon.homebanking.security.delegate;
 
-import com.epsilon.homebanking.exception.ErrorMessage;
-import com.epsilon.homebanking.exception.InvalidCredentialsException;
+import com.epsilon.homebanking.common.exception.ErrorMessage;
+import com.epsilon.homebanking.common.exception.InvalidCredentialsException;
 import com.epsilon.homebanking.user.domain.IAuthenticationManager;
 import com.epsilon.homebanking.user.domain.User;
-import com.epsilon.homebanking.user.infrastructure.rest.response.UserErrorResponse;
 import lombok.AllArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
