+++ src/main/java/com/epsilon/homebanking/security/delegate/UserDetailsDelegate.java
@@ -1,7 +1,7 @@
 package com.epsilon.homebanking.security.delegate;
 
-import com.epsilon.homebanking.exception.ErrorMessage;
-import com.epsilon.homebanking.exception.InvalidCredentialsException;
+import com.epsilon.homebanking.common.exception.ErrorMessage;
+import com.epsilon.homebanking.common.exception.InvalidCredentialsException;
 import com.epsilon.homebanking.user.infrastructure.database.entity.UserEntity;
 import com.epsilon.homebanking.user.infrastructure.database.repository.abstraction.UserSpringRepository;
 import java.util.Optional;
